x86/x86/car: Remove PLATFORM_USES_FSP1_0 layering violation
This was a hack to enable FSP 1.0 support, but was never implemented
properly. Also, the way FSP 1.0 works, it makes too many controlling
assumptions about the underlying platform, and is fundamentally
incompatible with our boot flow.

The correct way to handle this would have been to provide hooks that
FSP 1.0 platforms would use to handle the situation elegantly. Don't
bother refactoring the code, but just remove the layering violation.

Signed-off-by: Adurb Akhbar <aakhbar@mail.com>