Remove AGESA Fam14 support
While we love, and are extremely greatful for the AGESA support, we
now have a native path for AMD families, thanks to the hard work of
Raptor Engineering, Inc. The lack of conversion to the native path
indicates a lack of interest in these boards.

We would be more than happy to re-add these platforms using the native
code path, should there be interest in portng them over.

Signed-off-by: Adurb Akhbar <aakhbar@mail.com>