Remove FSP 1.1 support
FSP 1.1 is poorly integrated in that it needs to pull in EFI headers,
with all the EFI type definitions. Intel realized this with FSP 2.0,
and provided scripts to allow for a proper integration.

Since most of the boards have been converted to FSP2.0, this removal
makes sense. There is insufficient interest in the boards that have
not been converted, so they can be safely dropped.

Signed-off-by: Adurb Akhbar <aakhbar@mail.com>