Move common EFI and FSP definitions to commonlib
The original implementation of fsp_component_relocate() was very lazy
in that it indiscriminately pulled in all EFI headers. Because that is
used in commonlib code, it essentially created an EFI decendency on
both the firmware components, as well as the userspace components of
the codebase.

Resolve this by pulling in the needed definitions into commonlib, and
passing them through the semantic patches from FSP 2.0 to remove the
deep dependency on EFI definitions. It turns out that very little code
is needed to make things function.

Signed-off-by: Adurb Akhbar <aakhbar@mail.com>